EXSWHTEC-83 - Implement new and reimplement existing tests for hipMemAdvise
